+++ src/onegov.swissvotes/HISTORY.rst
@@ -7,6 +7,9 @@ Changelog
 - Adds datetime picker.
   [msom]
 
+- Allows to sort votes by turnout.
+  [msom]
+
 0.0.31 (2019-02-14)
 ~~~~~~~~~~~~~~~~~~~